pyproject.toml
[tool.poetry]
name = "djancial"
version = "0.1.0"
description = "Django social network"
authors = ["meeeeeeeh <alexanderkarpovich78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
Pillow = "^8.1.0"
psycopg2-binary = "^2.8.6"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.6.0"
incremental = "^17.5.0"
asgiref = "^3.3.1"
channels = "^3.0.3"
channels-redis = "^3.2.0"
aioredis = "^1.3.1"
asn1crypto = "^1.4.0"
async-timeout = "^3.0.1"
attrs = "^20.3.0"
autobahn = "^20.12.3"
Automat = "^20.2.0"
cffi = "^1.14.4"
constantly = "^15.1.0"
cryptography = "^3.3.1"
daphne = "^3.0.1"
hiredis = "^1.1.0"
hyperlink = "^21.0.0"
idna = "^3.1"
msgpack = "^1.0.2"
pycparser = "^2.20"
PyHamcrest = "^2.0.2"
pytz = "^2020.5"
six = "^1.15.0"
sqlparse = "^0.4.1"
Twisted = "^20.3.0"
txaio = "^20.12.1"
"zope.interface" = "^5.2.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
